moved damage color coding and labels to config.js for #2 & #5 final fix
