+++ src/common/vega-spec.js
@@ -1,9 +1,9 @@
 import { getData, renderVega } from "./mapd-connector";
 import { conv4326To900913 } from "./map-utils";
+import { getLabel, getColor } from "./config";
 import { updateMap } from "../components/map";
 import { updateCounterLabel } from '../components/counter-label';
 import { updateDamageChart } from '../components/damage-chart';
-import { getColor } from './damage-color-palette';
 
 export const createVegaSpec = ({map, endDateString, damageFilter}) => {
   // get map size
